requirements.txt
openai ==0.28
numpy
pandas
python-dotenv
